v0.3-SNAPSHOT : Refactoring
