+++ .gitignore
@@ -1,6 +1,7 @@
 .idea
 *.iml
 pom.xml.releaseBackup
+pom.xml.versionsBackup
 release.properties
 build
 .allure
