+++ client/src/test/java/io/github/jaqat/remoterobot/tests/config/TestsConfig.java
@@ -1,0 +1,7 @@
+package io.github.jaqat.remoterobot.tests.config;
+
+public class TestsConfig {
+
+    public final static int SCREEN_WIDTH = 1440;
+    public final static int SCREEN_HEIGHT = 900;
+}
