+++ client/src/test/java/io/github/jaqat/remoterobot/tests/enums/RemoteBrowserProvider.java
@@ -1,0 +1,5 @@
+package io.github.jaqat.remoterobot.tests.enums;
+
+public enum RemoteBrowserProvider {
+    GGR, SELENOID
+}
