+++ client/src/test/resources/tests/chrome/ColoredPage.png
