+++ client/src/test/resources/tests/chrome/PartitialColoredPage.png
