+++ client/src/test/resources/tests/firefox/PartitialColoredPage.png
