+++ commons/pom.xml
@@ -1,15 +1,44 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.github.jaqat.remoterobot</groupId>
+    <groupId>io.github.jaqat.remoterobot</groupId>
     <artifactId>commons</artifactId>
-    <version>0.1.6</version>
+    <version>0.3-SNAPSHOT</version>
 
-    <description>Remote Robot Commons</description>
+    <description>Remote Robot Server</description>
 
     <parent>
-        <groupId>com.github.jaqat.remoterobot</groupId>
+        <groupId>io.github.jaqat.remoterobot</groupId>
         <artifactId>parent</artifactId>
-        <version>0.1.6</version>
+        <version>0.3-SNAPSHOT</version>
     </parent>
+
+    <build>
+        <plugins>
+            <plugin>
+                <artifactId>maven-assembly-plugin</artifactId>
+                <configuration>
+                    <archive>
+                        <manifest>
+                            <mainClass>io.github.jaqat.remoterobot.server.Server</mainClass>
+                        </manifest>
+                    </archive>
+                    <descriptorRefs>
+                        <descriptorRef>jar-with-dependencies</descriptorRef>
+                    </descriptorRefs>
+                    <finalName>remote-robot-server</finalName>
+                    <appendAssemblyId>false</appendAssemblyId>
+                </configuration>
+                <executions>
+                    <execution>
+                        <id>make-assembly</id>
+                        <phase>package</phase>
+                        <goals>
+                            <goal>single</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
 
 </project>
