--- commons/src/main/java/com/github/jaqat/remoterobot/commons/config/Config.java
@@ -1,6 +1,0 @@
-package com.github.jaqat.remoterobot.commons.config;
-
-public class Config {
-
-    public static String REMOTE_ROBOT_SERVER_PORT = "5555";
-}
