--- commons/src/main/java/com/github/jaqat/remoterobot/commons/protocol/Operation.java
@@ -1,9 +1,0 @@
-package com.github.jaqat.remoterobot.commons.protocol;
-
-public enum Operation {
-    KEY_PRESS,
-    KEY_RELEASE,
-    MOUSE_CLICK,
-    GET_PIXEL_COLOR,
-    CAPTURE_SCREEN
-}
