+++ commons/src/main/java/io/github/jaqat/remoterobot/config/Config.java
@@ -1,0 +1,6 @@
+package io.github.jaqat.remoterobot.config;
+
+public class Config {
+
+    public static String REMOTE_ROBOT_SERVER_PORT = "5555";
+}
