+++ commons/src/main/java/io/github/jaqat/remoterobot/protocol/Operation.java
@@ -1,0 +1,9 @@
+package io.github.jaqat.remoterobot.protocol;
+
+public enum Operation {
+    KEY_PRESS,
+    KEY_RELEASE,
+    MOUSE_CLICK,
+    GET_PIXEL_COLOR,
+    CAPTURE_SCREEN
+}
