+++ commons/src/main/java/io/github/jaqat/remoterobot/protocol/Request.java
@@ -1,0 +1,24 @@
+package io.github.jaqat.remoterobot.protocol;
+
+import java.io.Serializable;
+import java.util.Arrays;
+import java.util.List;
+
+public class Request implements Serializable {
+
+    private Operation operation;
+    private List<Object> operands;
+
+    public Request(Operation operation, Object... operands) {
+        this.operation = operation;
+        this.operands = Arrays.asList(operands);
+    }
+
+    public Operation getOperation() {
+        return operation;
+    }
+
+    public List<Object> getOperands() {
+        return operands;
+    }
+}
