+++ docker-images/selenoid/build_and_deploy_image.sh
@@ -1,10 +1,9 @@
 #!/usr/bin/env bash
 
-#IMAGE=aafrikanov/remoterobot/selenoid_${ORIGINAL_SELENOID_IMAGE}:${VERSION}
-IMAGE=remoterobot/selenoid_${ORIGINAL_SELENOID_IMAGE}:${VERSION}
+IMAGE=remoterobot/selenoid_${ORIGINAL_SELENOID_IMAGE}:v${VERSION}
 
 echo "### Build '"${IMAGE}"' Docker image..."
 docker build -f docker-images/selenoid/Dockerfile --build-arg ORIGINAL_SELENOID_IMAGE=${ORIGINAL_SELENOID_IMAGE} -t ${IMAGE} .
 
 echo "### Deploy '"${IMAGE}"' Docker image..."
-#docker push ${IMAGE}
+docker push ${IMAGE}
