--- selenoid-utils/src/test/java/com/github/jaqat/remoterobot/selenoid/test/config/TestsConfig.java
@@ -1,7 +1,0 @@
-package com.github.jaqat.remoterobot.selenoid.test.config;
-
-public class TestsConfig {
-
-    public final static int SCREEN_WIDTH = 1440;
-    public final static int SCREEN_HEIGHT = 900;
-}
