--- selenoid-utils/src/test/java/com/github/jaqat/remoterobot/selenoid/test/enums/RemoteBrowserProvider.java
@@ -1,5 +1,0 @@
-package com.github.jaqat.remoterobot.selenoid.test.enums;
-
-public enum RemoteBrowserProvider {
-    GGR, SELENOID
-}
