--- selenoid-utils/src/test/resources/selenoid/config/browsers.json
@@ -1,24 +1,0 @@
-{
-  "firefox": {
-    "default": "64",
-    "versions": {
-      "64": {
-        "image": "remoterobot/selenoid_firefox_64.0:0.1.6",
-        "port": "4444",
-        "path": "/wd/hub",
-        "publishAllPorts": true
-      }
-    }
-  },
-  "chrome": {
-    "default": "70",
-    "versions": {
-      "70": {
-        "image": "remoterobot/selenoid_chrome_71.0:0.1.6",
-        "port": "4444",
-        "path": "/",
-        "publishAllPorts": true
-      }
-    }
-  }
-}
