+++ server/pom.xml
@@ -1,22 +1,28 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.github.jaqat.remoterobot</groupId>
+    <groupId>io.github.jaqat.remoterobot</groupId>
     <artifactId>server</artifactId>
-    <version>0.1.6</version>
+    <version>0.3-SNAPSHOT</version>
 
     <description>Remote Robot Server</description>
 
     <parent>
-        <groupId>com.github.jaqat.remoterobot</groupId>
+        <groupId>io.github.jaqat.remoterobot</groupId>
         <artifactId>parent</artifactId>
-        <version>0.1.6</version>
+        <version>0.3-SNAPSHOT</version>
     </parent>
 
     <dependencies>
         <dependency>
-            <groupId>com.github.jaqat.remoterobot</groupId>
+            <groupId>io.github.jaqat.remoterobot</groupId>
             <artifactId>commons</artifactId>
-            <version>0.1.6</version>
+            <version>0.3-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
+            <groupId>io.github.jaqat.remoterobot</groupId>
+            <artifactId>client</artifactId>
+            <version>0.3-SNAPSHOT</version>
         </dependency>
     </dependencies>
 
@@ -27,7 +33,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>com.github.jaqat.remoterobot.server.Server</mainClass>
+                            <mainClass>io.github.jaqat.remoterobot.server.Server</mainClass>
                         </manifest>
                     </archive>
                     <descriptorRefs>
@@ -38,8 +44,8 @@
                 </configuration>
                 <executions>
                     <execution>
-                        <id>make-assembly</id> <!-- this is used for inheritance merges -->
-                        <phase>package</phase> <!-- bind to the packaging phase -->
+                        <id>make-assembly</id>
+                        <phase>package</phase>
                         <goals>
                             <goal>single</goal>
                         </goals>
