--- server/src/main/java/com/github/jaqat/remoterobot/server/utils/OperationExecution.java
@@ -1,10 +1,0 @@
-package com.github.jaqat.remoterobot.server.utils;
-
-@FunctionalInterface
-public interface OperationExecution {
-
-    /*
-     Execute actions that can return some result object
-     */
-    Object execute() throws Exception;
-}
