+++ server/src/main/java/io/github/jaqat/remoterobot/server/utils/MouseUtil.java
@@ -1,0 +1,23 @@
+package io.github.jaqat.remoterobot.server.utils;
+
+
+import io.github.jaqat.remoterobot.protocol.Response;
+
+import static io.github.jaqat.remoterobot.protocol.Operation.MOUSE_CLICK;
+
+public class MouseUtil extends JavaRobotUtil {
+
+    public Response mouseClick(int x, int y, int buttonMask) {
+        return executeOperation(
+                MOUSE_CLICK,
+                () -> {
+                    robot.mouseMove(x, y);
+                    robot.mousePress(buttonMask);
+                    robot.mouseRelease(buttonMask);
+                    return null;
+                },
+                (result) -> String.format("Clicked to coordinates [%d : %d]", x, y),
+                "Unable to mouse click."
+        );
+    }
+}
