+++ server/src/main/java/io/github/jaqat/remoterobot/server/utils/OperationExecution.java
@@ -1,0 +1,10 @@
+package io.github.jaqat.remoterobot.server.utils;
+
+@FunctionalInterface
+public interface OperationExecution {
+
+    /*
+     Execute actions that can return some result object
+     */
+    Object execute() throws Exception;
+}
